requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pyyaml>=6.0.1
numpy>=1.26.0
opencv-python>=4.10.0
pillow>=10.0.0
tqdm>=4.66.0
scikit-image>=0.24.0
flask>=3.0.0
tensorboard>=2.16.0
